+++ tests/testthat/test-hyperlink.R
@@ -1,12 +1,6 @@
 
-context("Hyperlinks")
-
-op <- options()
-on.exit(options(op))
-options(crayon.enabled = TRUE)
-options(crayon.hyperlink = TRUE)
-
 test_that("hyperlinks", {
+  withr::local_options(crayon.hyperlink = TRUE)
   expect_equal(
     hyperlink("foo", "https://bar"),
     "\033]8;;https://bar\afoo\033]8;;\a"