+++ tests/testthat/test-vectors.r
@@ -1,15 +1,9 @@
 
-context("Styling of character vectors")
-
-op <- options()
-on.exit(options(op))
-options(crayon.enabled = TRUE)
-
 foobar <- c("foo", "bar")
 bigyo <- c("bi", "gyo")
 
 test_that("Coloring and highlighting works", {
-
+  local_colors()
   expect_equal(underline(foobar),
                c('\u001b[4mfoo\u001b[24m', '\u001b[4mbar\u001b[24m'))
   expect_equal(red(foobar),
@@ -20,7 +14,7 @@ test_that("Coloring and highlighting works", {
 })
 
 test_that("Applying multiple styles at once works", {
-
+  local_colors()
   expect_equal(red$bgGreen$underline(foobar),
                c('\u001b[31m\u001b[42m\u001b[4mfoo\u001b[24m\u001b[49m\u001b[39m',
                  '\u001b[31m\u001b[42m\u001b[4mbar\u001b[24m\u001b[49m\u001b[39m'))
@@ -30,7 +24,7 @@ test_that("Applying multiple styles at once works", {
 })
 
 test_that("Nested styles are supported", {
-
+  local_colors()
   expect_equal(
     red(foobar %+% underline$bgBlue(bigyo) %+% '!'),
     c('\u001b[31mfoo\u001b[4m\u001b[44mbi\u001b[49m\u001b[24m!\u001b[39m',
@@ -38,7 +32,7 @@ test_that("Nested styles are supported", {
 })
 
 test_that("Nested styles of the same type are supported", {
-
+  local_colors()
   aA <- c("a", "A")
   bB <- c("b", "B")
   cC <- c("c", "C")
@@ -49,7 +43,7 @@ test_that("Nested styles of the same type are supported", {
 })
 
 test_that("Reset all styles", {
-
+  local_colors()
   expect_equal(reset(red$bgGreen$underline(foobar) %+% foobar),
 
 c("\033[0m\033[31m\033[42m\033[4mfoo\033[24m\033[49m\033[39mfoo\033[0m\033[22m\033[23m\033[24m\033[27m\033[28m\033[29m\033[39m\033[49m", "\033[0m\033[31m\033[42m\033[4mbar\033[24m\033[49m\033[39mbar\033[0m\033[22m\033[23m\033[24m\033[27m\033[28m\033[29m\033[39m\033[49m")
@@ -57,7 +51,7 @@ c("\033[0m\033[31m\033[42m\033[4mfoo\033[24m\033[49m\033[39mfoo\033[0m\033[22m\0
 })
 
 test_that("Variable number of arguments", {
-
+  local_colors()
   expect_equal(red(foobar, 'bar'),
                c('\u001b[31mfoo bar\u001b[39m',
                  '\u001b[31mbar bar\u001b[39m'))